Allow second sync committee message to propagate if first one is stale
The sync committee message signature does not sign over the slot, making
it possible for stale messages to be adjusted and replayed, possibly
preventing new sync committee messages to propagate. If such a message
for an older head is received, allow a second message to propagate that
refers to the current head as selected by fork choice, overriding the
previous one.

Also add a couple sanity checks, as in, disallowing messages referring
to unknown or invalid blocks to spread, in line with existing rules for
blob sidecars, parent block roots and attestations.