Use 32-bit random value in validator filter
Currently, we use an 8-bit random byte to perform an effective balance filter.
This was used to penalize validators with effective balances less than 32 ETH.
But now that we have raised the maximum effective balance, this mechanism plays
a more important role in validator selection.

One might expect a validator with a greater effective balance to have a greater
chance of being selected as proposer or sync committee participant, but no.
This is because the random byte can only represent 256 "buckets" of effective
balance. So for example, currently, a validator with 33 ETH has the same
likelihood of being selected as a validator with 40 ETH, because these
validators are in the same bucket.

The solution is to use a bigger random value to allow more precision. I propose
we use a 32-bit random value instead which provides more than enough precision.
But this change requires we use "effective balance increments" instead of Gwei
in the calculation because the latter could cause an integer overflow when
multiplying the values together.