Make necessary changes for EIP7594 spec
Previous to this commit, the max blob count for EIP7594 was 8. Now that we have
raised the target/max to 6/9 for Electra, the max blob count for EIP7594 no
longer makes sense. I've raised the target/max for it to 9/12 as a placeholder.
This also serves as an example of what changes are necessary to increase the
blob count in the future.