Clarify gossip limits
In the gossip specification, the `GOSSIP_MAX_SIZE` constant is specified
for the uncompressed payload size in the gossipsub message.

This PR clarifies how this limit applies to the various fields of the
gossipsub message and provides additional limits derived from it that
allow clients to more aggressively discard messages.

In particular, clients are allowed to impose more strict limits on
topics such as attestation and aggregates - an `Attestation` for example
takes no more than `~228` bytes (to be verified!), far below the 10mb
limit, though implicitly clients should already see these limits imposed
as rejections by their SSZ decoder - this clarification mainly
highlights the possibilty to perform this check earlier in the process.