Simplify build system for computed constants
When type definitions depend on preset values, those preset values have
to be manually specified in `hardcoded_custom_type_dep_constants` in a
manual and error prone way. By using heuristics similar to other parts
of the spec builder, that can be automated so that computed constants
are automatically generated _after_ preset values are emitted. This also
makes it clearer which constants are pure, and which constants depend on
specific preset values being used.