Move inclusion list specs to new `inclusion-list.md` file (#4446)
This PR adds `inclusion-list.md` as
[suggested](https://github.com/ethereum/consensus-specs/pull/4351#pullrequestreview-3023153485)
in the previous PR. `inclusion-list.md` encapsulates
`InclusionListStore` and inclusion list handling logic. Fork-choice
store interacts with IL store via functions defined in
`inclusion-list.md`, leveraging this abstraction.

This can be considered as adding another module that is responsible of
inclusion list. To represent some of the relationship between the
fork-choice module and others in a simple diagram, it would look like
this: [Beacon chain] — [Fork-choice] — [Inclusion list].