Make commitment deduplication in `verify_cell_kzg_proof_batch` deterministic (#4447)
We found a small issue w.r.t. the deduplication of commitments in
`verify_cell_kzg_proof_batch`:

```python
    # Create the list of deduplicated commitments we are dealing with
    deduplicated_commitments = [
        bytes_to_kzg_commitment(commitment_bytes) for commitment_bytes in set(commitments_bytes)
    ]
```

In short, using `set` to remove the duplicates is wrong as sets are
unordered objects:

```python
>>> a = [1, 3, 2, 1]
>>> list(set(a))
[1, 2, 3]
```

Because of that, between runs, `deduplicated_commitments` will be
different and so `commitment_indices`. The issue here is that the
Fiat-Shamir challenge is computed by taking this arguments and hashing
them together:

```python
def verify_cell_kzg_proof_batch_impl(
    commitments: Sequence[KZGCommitment],
    commitment_indices: Sequence[CommitmentIndex],
    cell_indices: Sequence[CellIndex],
    cosets_evals: Sequence[CosetEvals],
    proofs: Sequence[KZGProof],
) -> bool:

    ...

    # Step 1: Compute a challenge r and its powers r^0, ..., r^{num_cells-1}
    r = compute_verify_cell_kzg_proof_batch_challenge(
        commitments, commitment_indices, cell_indices, cosets_evals, proofs
    )

   ...
```

so, if between runs they change due to the non-deterministic
deduplication, so does the challenge. This breaks Fiat-Shamir because if
the prover generates a challenge $c_i$ and the verifier generates $c_j$
so that $c_i != c_j$ for the same set of inputs $S$, then proofs won't
be able to be verified, even if "correct".

Right now this is not an issue as this is used in PeerDAS as an
optimization, not like the usual two-parties setup we find in ZK.
However, for correctness, this should be addressed, as the specs are
"formally" wrong as of now.