Make several updates to EIP-7732 spec (#4438)
This PR updates EIP-7732 in several places. It should be read commit by
commit. The list of commits is as follows

- Add a Builder withdrawal prefix `0x03` which is a compounding
validator, allow any validator to self-build for a zero value, but only
0x03 validators to build for others.
- Remove all PTC reward logic and keep the beacon committee as pre-7732.
- Add slot to the execution payload envelope (makes implementation
easier)
- Remove the payload status enumeration and simply use a boolean in the
PTCAttestation object.
- Update the payment mechanism from a direct transfer of validator
balance to a withdrawal to a specific execution layer address (a fee
recipient).
- Use the attestation data index field to signal payload availability in
attestations.
- Fix all tests

This PR does not update the forkchoice logic, that will come on a
separate PR perhaps from @fradamt