In `get_custody_groups`, skip computation if all groups are custodied (#4459)
A simple optimization which doesn't change anything, but worth
mentioning because not all clients do this. For supernodes which custody
everything, we can skip computing the list of custody groups and return
the full range of groups. Given that there is quite a bit of hashing
involved, this isn't exactly free.

Thanks to @nalepae for suggesting this.