Fast test_full_attestations_one_validator_one_gwei (#4462)
Fixes https://github.com/ethereum/consensus-specs/issues/4318 

Now it takes seconds.

Before merging we need to consider:
 + With this version the one gwei validator is not the only validator.
+ As one gwei makes it less likely to be chosen for the committee, with
a different pseudo-random seed this test could fail.

The previous version would never fail because the one gwei validator was
the only one, so it was warrantied to be part of the committee. However,
the previous version could take arbitrary time depending on the
pseudo-random seed.

So this approach has its pros and cons with respect to previous version.