Update EIP-7732:  Add boolean to PayloadAttestationData (#4473)
This is just a placeholder EIP to have the right data structure. This
change should entail changes in the following places but those can only
be made after rebasing on top of Fulu and #3510

- The honest validator guide should include two separate deadlines for
the PTC, on payload available and data available.
- The honest validator guide should give instructions on how to set the
new data available flag to true if the data was available at the last
deadline and set the payload available flag if the payload was available
at the first deadline
- Forkchoice should update `get_proposer_head` with the payload and data
timeliness flags.