Make NUMBER_OF_COLUMNS a preset (#4485)
[As discussed on
discord](https://discord.com/channels/595666850260713488/1401858592264552501),
`NUMBER_OF_COLUMNS` should be a preset, not a configuration value. It's
technically an alias of `CELLS_PER_EXT_BLOB` (preset) which is used in
KZG libraries. I've also added the latter to the preset files as well so
that it is more clear that it's an alias.