Make PTC selection balance-weighted (#4488)
1. The first commit changes the sampling logic in `get_ptc` to be
balance weighted, for security reasons (see discord discussion starting
[here](https://discord.com/channels/595666850260713488/874767108809031740/1400916536197251082)).
Unlike for sync-committees and proposers, it doesn't use a new shuffle
to select candidate indices to which a balance-weighted acceptance
probability is applied. Instead, to minimize computational load, it just
reuses the existing committee shuffle, starting from the committees of
the PTC's `slot` and moving forward from there (wrapping around if
necessary).
2. The second commit refactors all functions that do balance weighted
sampling, through helpers `compute_balance_weighted_selection` and
`compute_balance_weighted_acceptance`:
- `compute_balance_weighted_acceptance` takes care of the
balance-weighted acceptance (whether a candidate index is accepted or
not, with probability proportional to its balance)
- `compute_balance_weighted_selection` takes care of the whole selection
process, going over the given `indices` and selecting them with
balance-weighted probability until the required number has been
selected. It has a parameter `shuffle_indices` which indicates whether
`indices` should be shuffled or not. This is `False` for the PTC, since
we are just reusing the existing shuffle
3. The other commits remove `state.latest_full_slot`, which is unused
after the introduction of `state.execution_payload_availability`, and
fixes a small mistake in `fork-choice.md`

If people think the refactor is unnecessary, we could just go with the
first commit. On the one end, the refactor might make it a bit more
annoying to change the selection logic for just some of these in the
future. On the other hand, there is currently a lot of code duplication
and lack of readability, and `get_ptc` is in particular quite hard to
understand because there's a lot going on in it.

Another option could be a smaller refactor with just
`compute_balance_weighted_acceptance`, which should be much less likely
to be customized by different selection processes (whereas
`compute_balance_weighted_selection` already has some customization with
`shuffle_indices`, and there's other possibilities like fixed size or
not)