eip7732: do not deduplicate `indexed_payload_attestations` (#4496)
`attesting_indices` for a payload attestation should be able to contain
duplicate indices if the PTC does (which it can, as of #4488), to verify
an aggregate signature/pubkey that contains some signature/pubkey
multiple times. This is also the behavior for sync committees, as
opposed to including a single signature/pubkey even for duplicate
indices, and having to deal with duplication more explicitly (e.g.
`aggregation_bits` could be less than `PTC_SIZE`, counting would have to
be with multiplicity).

Tests are fixed by making `prepare_signed_payload_attestation` deal with
duplicates in `attesting_indices` when setting `aggregation_bits`, by
enforcing 1 index = 1 bit set (if a validator index appears `n` times in
the PTC but only `m < n` in `attesting_indices`, only `m` bits are set)