Include pre-releases in release drafter & update labels (#4518)
This PR sets `include-pre-releases` to true and explicitly specifies
that `master` is the branch that tags are created from. For some reason,
release drafter is including old PR lines in the release notes which
I've had to manually delete. It's confused about what the previous tag
is or something. I'm hoping these changes fix it.

This PR also updates the labels so that `eip7594` and `eip7732` PRs are
included in the Fulu/Gloas sections, respectively. I don't want there to
be a disconnect here now. These should be under the fork section, not
the feature section.