Ensure cell indices are sorted in `recover_cells_and_kzg_proofs` (#4519)
This PR rejects shuffled inputs (cell indices & cells) that are provided
to `recover_cells_and_kzg_proofs`. Previously, we assumed that the
inputs were in the correct order; now we enforce it. I've added two
reference tests to ensure implementations handle this situation
appropriately.

Thanks to @asanso for pointing this out.