Do not run KZG compute challenge tests in the collection phase (#4546)
The alpha.6 release stalled in the reference test generation step
because of this. When adding these, I forgot to use an input getter
function, so it was computing the blob's commitment in the test
collection phase. For some reason, this caused latest usage of
`arkworks_G1.multiexp_unchecked` to completely halt. I think it has
something to do with mixing up Rust's `rayon` and Python's threading
[here](https://github.com/crate-crypto/py-arkworks-bls12381/blob/ce5f2cedd53a3cf7e1f9e35921b2fbbf43d55674/src/wrapper.rs#L87).