Add note about fee recipient (#4565)
Fixes #4560 

I think we need to add a builder API file completely specified now
in-protocol to specify how validators can contact directly builders to
request a bid. Perhaps this can be added to the `validator.md` file
directly or we add a new file for this. Whenever this communication is
specified we need to add a full description on how builders get this fee
recipient and how validators specify it. For the time being, this PR
just adds the note to the validator.md.