Use skips instead of asserting when an SSZ test is cancelled
SSZ tests are scheduled asynchronously since #4574, but those that are
later determined to be inapplicable should be properly cancelled by
raising SkippedTest instead of aborting test generation.

Notably, this is about tests where a valid object is serialized, the
serialized data is subsequently modified and parsed again. As the
modified data may still be valid depending on the structure, such
tests have to be skipped as they do not denote invalid data.