eip7732: remove header and inclusion proof from data column sidecar (#4527)
This PR removes the header and inclusion proof from the data column
sidecar, based on the assumption that the beacon block containing the
bid and KZG commitment root is received before the data column sidecar.
The reason these fields exist today is that sidecars and blocks are
broadcast at the same time.