Only create ssz_static tests in generator mode (#4596)
This should fix the crazy slow tests. It seems that `_create_test_case`
is just crazy slow. When we have more time, we should investigate why
this is exactly. I suspect there are just so many test cases that it
takes forever.

My "dumb solution" is to skip calling `_create_test_cases` outside of
generator mode.

Related to:

* #4573
* #4584