Make RNG deterministic again in ssz_generic runner
With #4574, the same RNG instance was captured by the various `the_test`
functions. When they were subsequently executed in parallel, this led to
non-deterministic tests. However, #4639 prefers reproducible tests.

Instead of running tests in parallel on a shared RNG instance, the RNG
instance is now used during the single-threaded test collection stage
to produce a deterministic order of seeds, which is then used to create
a per-test RNG instance that is not shared with any other tests, hence
not being subject to non-determinism from parallel execution.