Ensure data column sidecars respect the blob limit (#4650)
The gossip validation for data column sidecars is missing a check which
ensures the number of blob KZG commitments in the sidecar respects the
max blobs per block limit. This would allow a malicious proposer to
broadcast sidecars with a bunch of data to verify (slow) and it would
still be propagated causing other nodes to waste resources on this too.
I'm not very concerned with this as a DoS, but we should add a check for
this.

Instead of adding a line to the [gossip verification
checks](https://github.com/ethereum/consensus-specs/blob/master/specs/fulu/p2p-interface.md#data_column_sidecar_subnet_id)
I thought it would be better to put the check in
`verify_data_column_sidecar`. Code is better than an english sentence 🙂

Edit: This came from a report in the fusaka audit competition.
Thank you!