Restart workers periodically to prevent memory accumulation (#4666)
I'm pretty sure the reference test generation working functions are
accumulating memory and this is why the runner is running out of memory.
I tested locally and found that this did make a difference. I will merge
this & see if it solves our issue.

I think switching to pytest might solve the underlying issue, not sure.
I believe this issue popped up now because etan's PR added many new SSZ
tests and that pushed it over the limit.