Maybe fix release push step (#4675)
The spec release is still failing to push the spec tests. Not sure why,
but I think removing the `--depth=1` argument will fix it. It appears
that it's referencing an object which it doesn't have. This used to
work, not sure what changed. Maybe it had something to do with
deterministic builds.

I've also added a step to include the release tarballs in the
consensus-specs release.