Do not push files to consensus-spec-tests (#4678)
I'm not sure how to fix this issue in the release action. Something is
wrong with the runner or with git lfs.

For this reason, this PR updates the release action to **not** push to
the consensus-spec-tests repo. We can try to deprecate this repo. Spec
tests are now included in the consensus-specs release. This should be a
fairly small change for clients.