Set `block_hash` in the latest bid during Gloas state upgrade (#4739)
When processing the first Gloas block, we check if
`is_parent_block_full` in order to process withdrawals. This will be
false for the first block and lead to failed withdrawals root checking
during payload processing. This PR implements the simplest solution by
setting the `block_hash` in the bid during the state upgrade.